web-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary,pool]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
pydantic
pydantic-settings
email-validator
